# app/app/docs/swagger.yaml
swagger: "2.0"
info:
  description: "API documentation for AE2I application"
  version: "1.0.0"
  title: "AE2I API"
host: "localhost:5000"
basePath: "/api"
schemes:
  - "http"
securityDefinitions:
  BearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"
    description: "Ajouter 'Bearer <votre_token>' dans le champ Authorization pour authentifier les requêtes."
...
paths:
  /users:
    get:
      tags:
        - "users"
      summary: "Get all users"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "A list of users"
        500:
          description: "Internal server error"
  /users/{email}:
    get:
      tags:
        - "users"
      summary: "Get user by email"
      security:
        - BearerAuth: []
      parameters:
        - name: "email"
          in: "path"
          required: true
          type: "string"
      responses:
        200:
          description: "User found"
        404:
          description: "User not found"
        500:
          description: "Internal server error"
  /users/{user_id}:
    put:
      tags:
        - "users"
      summary: "Update user"
      parameters:
        - name: "user_id"
          in: "path"
          required: true
          type: "string"
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              email:
                type: "string"
              firstName:
                type: "string"
              lastName:
                type: "string"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "User updated successfully"
        404:
          description: "User not found"
        500:
          description: "Internal server error"
    delete:
      tags:
        - "users"
      summary: "Delete user"
      parameters:
        - name: "user_id"
          in: "path"
          required: true
          type: "string"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "User deleted successfully"
        404:
          description: "User not found"
        500:
          description: "Internal server error"
  /users/{user_id}/cart:
    post:
      tags:
        - "users"
      summary: "Add item to cart"
      parameters:
        - name: "user_id"
          in: "path"
          required: true
          type: "string"
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              productId:
                type: "string"
              quantity:
                type: "integer"
              price:
                type: "number"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Item added to cart"
        500:
          description: "Internal server error"
    get:
      tags:
        - "users"
      summary: "Get user cart"
      parameters:
        - name: "user_id"
          in: "path"
          required: true
          type: "string"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "User cart retrieved"
        500:
          description: "Internal server error"
    delete:
      tags:
        - "users"
      summary: "Flush user cart"
      parameters:
        - name: "user_id"
          in: "path"
          required: true
          type: "string"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Cart flushed"
        500:
          description: "Internal server error"
  /users/{user_id}/cart/{item_id}:
    delete:
      tags:
        - "users"
      summary: "Delete item from cart"
      parameters:
        - name: "user_id"
          in: "path"
          required: true
          type: "string"
        - name: "item_id"
          in: "path"
          required: true
          type: "string"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Item deleted from cart"
        500:
          description: "Internal server error"
...
  /posts:
    get:
      tags:
        - "posts"
      summary: "Get all posts"
      responses:
        200:
          description: "A list of posts"
        500:
          description: "Internal server error"
    post:
      tags:
        - "posts"
      summary: "Create a new post"
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              title:
                type: "string"
              description:
                type: "string"
              link:
                type: "string"
      security:
        - BearerAuth: []
      responses:
        201:
          description: "Post created successfully"
        500:
          description: "Internal server error"
  /posts/{post_id}:
    get:
      tags:
        - "posts"
      summary: "Get post by ID"
      parameters:
        - name: "post_id"
          in: "path"
          required: true
          type: "string"
      responses:
        200:
          description: "Post found"
        404:
          description: "Post not found"
        500:
          description: "Internal server error"
    put:
      tags:
        - "posts"
      summary: "Update post"
      parameters:
        - name: "post_id"
          in: "path"
          required: true
          type: "string"
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              title:
                type: "string"
              description:
                type: "string"
              link:
                type: "string"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Post updated successfully"
        404:
          description: "Post not found"
        500:
          description: "Internal server error"
    delete:
      tags:
        - "posts"
      summary: "Delete post"
      parameters:
        - name: "post_id"
          in: "path"
          required: true
          type: "string"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Post deleted successfully"
        404:
          description: "Post not found"
        500:
          description: "Internal server error"
...
  /polls:
    get:
      tags:
        - "polls"
      summary: "Get all polls"
      responses:
        200:
          description: "A list of polls"
        500:
          description: "Internal server error"
    post:
      tags:
        - "polls"
      summary: "Create a new poll"
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              title:
                type: "string"
              description:
                type: "string"
              choices:
                type: "array"
                items:
                  type: "object"
                  properties:
                    choice_text:
                      type: "string"
      security:
        - BearerAuth: []
      responses:
        201:
          description: "Poll created successfully"
        500:
          description: "Internal server error"
...
  /polls/{poll_id}/responses:
    get:
      tags:
        - "polls"
      summary: "Get responses by poll ID"
      parameters:
        - name: "poll_id"
          in: "path"
          required: true
          type: "string"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Responses found"
        500:
          description: "Internal server error"
    post:
      tags:
        - "polls"
      summary: "Create response"
      parameters:
        - name: "poll_id"
          in: "path"
          required: true
          type: "string"
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              choice:
                type: "string"
      security:
        - BearerAuth: []
      responses:
        201:
          description: "Response created successfully"
        500:
          description: "Internal server error"
...
  /articles:
    get:
      tags:
        - "articles"
      summary: "Get all articles"
      responses:
        200:
          description: "A list of articles"
        500:
          description: "Internal server error"
    post:
      tags:
        - "articles"
      summary: "Create a new article"
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              title:
                type: "string"
              description:
                type: "string"
              price:
                type: "number"
              photo:
                type: "string"
      security:
        - BearerAuth: []
      responses:
        201:
          description: "Article created successfully"
        500:
          description: "Internal server error"
  /articles/{article_id}:
    get:
      tags:
        - "articles"
      summary: "Get article by ID"
      parameters:
        - name: "article_id"
          in: "path"
          required: true
          type: "string"
      responses:
        200:
          description: "Article found"
        404:
          description: "Article not found"
        500:
          description: "Internal server error"
    put:
      tags:
        - "articles"
      summary: "Update article"
      parameters:
        - name: "article_id"
          in: "path"
          required: true
          type: "string"
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              title:
                type: "string"
              description:
                type: "string"
              price:
                type: "number"
              photo:
                type: "string"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Article updated successfully"
        404:
          description: "Article not found"
        500:
          description: "Internal server error"
    delete:
      tags:
        - "articles"
      summary: "Delete article"
      parameters:
        - name: "article_id"
          in: "path"
          required: true
          type: "string"
      security:
        - BearerAuth: []
      responses:
        200:
          description: "Article deleted successfully"
        404:
          description: "Article not found"
        500:
          description: "Internal server error"
...